+++ examples/with_custom_ui.rs
@@ -163,11 +163,11 @@ async fn progress_bar_task(file_size: u64, mut info_rx: broadcast::Receiver<Down
 #[tokio::main]
 async fn main() {
     // 初始化 tracing（生产默认 info，调试默认 debug；RUST_LOG 优先）
-    simple_downloader::trace::init_tracing();
+    // simple_downloader::trace::init_tracing();
     // 1. --- 配置 ---
 
-    let url = "https://dlied4.myapp.com/myapp/1104466820/cos.release-40109/10040714_com.tencent.tmgp.sgame_a2480356_8.2.1.9_F0BvnI.apk";
-    let output_path = "10040714_com.tencent.tmgp.sgame_a2480356_8.2.1.9_F0BvnI.apk";
+    let url = "https://www.modelscope.cn/models/Tencent-Hunyuan/Hy-MT2-1.8B-GGUF/resolve/master/Hy-MT2-1.8B-Q6_K.gguf";
+    let output_path = "Hy-MT2-1.8B-Q6_K.gguf";
 
     let workers = 16; // 最大并发数
     let update_interval = 0.2; // 更新间隔为 0.2 秒